# backend/indexer/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/content": {
            "get": {
                "description": "Get list of tracked content with violation stats",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "content"
                ],
                "summary": "List content",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search by title",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by Kinopoisk ID",
                        "name": "kinopoisk_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by IMDB ID",
                        "name": "imdb_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by MAL ID",
                        "name": "mal_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by Shikimori ID",
                        "name": "shikimori_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by MyDramaList ID",
                        "name": "mydramalist_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by violations presence (true/false)",
                        "name": "has_violations",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "violations_count",
                            "created_at"
                        ],
                        "type": "string",
                        "default": "violations_count",
                        "description": "Sort by field",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "Sort order",
                        "name": "sort_order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.ListContentResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Add content to track for violations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "content"
                ],
                "summary": "Create content",
                "parameters": [
                    {
                        "description": "Content data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.CreateContentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.ContentWithStats"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/content/batch": {
            "post": {
                "description": "Add multiple content items to track for violations in a single batch request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "content"
                ],
                "summary": "Create multiple content items",
                "parameters": [
                    {
                        "description": "Content items data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.CreateContentBatchRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.CreateContentBatchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/content/check-violations": {
            "post": {
                "description": "Refresh violation stats for selected content items",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "content"
                ],
                "summary": "Check violations for content",
                "parameters": [
                    {
                        "description": "Content IDs to check",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.CheckViolationsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.CheckViolationsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/content/delete": {
            "post": {
                "description": "Remove multiple content items from tracking",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "content"
                ],
                "summary": "Delete multiple content items",
                "parameters": [
                    {
                        "description": "Content IDs to delete",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.DeleteContentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.DeleteContentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/content/{id}": {
            "get": {
                "description": "Get content details with violation stats",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "content"
                ],
                "summary": "Get content by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Content ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.ContentWithStats"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove content from tracking",
                "tags": [
                    "content"
                ],
                "summary": "Delete content",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Content ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/content/{id}/violations": {
            "get": {
                "description": "Get list of pages where content was found",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "content"
                ],
                "summary": "Get violations for content",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Content ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.ListViolationsResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/content/{id}/violations/export": {
            "get": {
                "description": "Export all violations for content to CSV file",
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "content"
                ],
                "summary": "Export violations to CSV",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Content ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/indexer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "indexer_internal_handler.ContentWithStats": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imdb_id": {
                    "type": "string"
                },
                "kinopoisk_id": {
                    "type": "string"
                },
                "mal_id": {
                    "type": "string"
                },
                "mydramalist_id": {
                    "type": "string"
                },
                "original_title": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "shikimori_id": {
                    "type": "string"
                },
                "sites_count": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "violations_count": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "indexer_internal_handler.CreateContentRequest": {
            "type": "object",
            "properties": {
                "imdb_id": {
                    "type": "string"
                },
                "kinopoisk_id": {
                    "type": "string"
                },
                "mal_id": {
                    "type": "string"
                },
                "mydramalist_id": {
                    "type": "string"
                },
                "original_title": {
                    "type": "string"
                },
                "shikimori_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "indexer_internal_handler.ListContentResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/indexer_internal_handler.ContentWithStats"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    }
}